requirements.txt
flask
supabase
gunicorn
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
sqlalchemy-libsql
Flask-Login
Flask-WTF
boto3
email_validator
psycopg2
APScheduler
python-slugify